refactor(portfolio) : 포트폴리오 서비스 디렉터리 추가(#19)
service 디렉터리 하위에 portfolio 디렉터리 추가
포트폴리오 페이지에 누락된 img alt 속성 추가
